Refactor Storybook mocks
